Fix stamp.r
With the switch to 'sf' from 'sp', stamp(...) broke when a DISA or GENR or STBL event LEV1 didn't occur. Any modification to the sf objects with no data frames (i.e., polygonal geometry) would result in an error. I added some if statements to see if a data frame was associated with the gd1/gd2/pI objects before modifying the data frames and combining them.