Fix stamp.direction.r
Two fixes here:
(1) code was failing when more than one object was in a group but both objects were in the second period. Fix - check if the group contains objects during both time periods.
(2) st_geom_azimuth(...) was spitting out a non-numeric datatype, which was producing an error when the code tried to make the negative values on the 0-360 scale.